extract material buffer creation to shared filed
