+++ src/renderer/glsl/chunks/materialBuffer.glsl
@@ -1,0 +1,32 @@
+export default `
+
+uniform Materials {
+  vec4 colorAndMaterialType[NUM_MATERIALS];
+  vec4 roughnessMetalnessNormalScale[NUM_MATERIALS];
+
+  #if defined(NUM_DIFFUSE_MAPS) || defined(NUM_NORMAL_MAPS) || defined(NUM_PBR_MAPS)
+    ivec4 diffuseNormalRoughnessMetalnessMapIndex[NUM_MATERIALS];
+  #endif
+
+  #if defined(NUM_DIFFUSE_MAPS) || defined(NUM_NORMAL_MAPS)
+    vec4 diffuseNormalMapSize[NUM_DIFFUSE_NORMAL_MAPS];
+  #endif
+
+  #if defined(NUM_PBR_MAPS)
+    vec2 pbrMapSize[NUM_PBR_MAPS];
+  #endif
+} materials;
+
+#ifdef NUM_DIFFUSE_MAPS
+  uniform mediump sampler2DArray diffuseMap;
+#endif
+
+#ifdef NUM_NORMAL_MAPS
+  uniform mediump sampler2DArray normalMap;
+#endif
+
+#ifdef NUM_PBR_MAPS
+  uniform mediump sampler2DArray pbrMap;
+#endif
+
+`;
